requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
alembic==1.13.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
async-timeout==5.0.1
attrs==25.3.0
bcrypt==4.0.1
billiard==4.2.2
celery==5.3.6
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==46.0.1
databases==0.9.0
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.1
email_validator==2.2.0
eval-type-backport==0.1.3
exceptiongroup==1.3.0
fastapi==0.111.0
fastapi-cli==0.0.7
filelock==3.19.1
frozenlist==1.7.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.27.0
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
kombu==5.5.4
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.6.4
numpy==2.2.6
openai==1.30.1
orjson==3.11.3
packaging>=21,<25
passlib==1.7.4
pillow==10.4.0
pluggy==1.6.0
prompt_toolkit==3.0.52
propcache==0.3.2
pyarrow==21.0.0
pyasn1==0.6.1
pycparser==2.23
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.2.2
pytest-asyncio==0.23.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.3
redis==5.0.8
requests==2.32.3
rich==14.1.0
rich-toolkit==0.15.1
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.34
starlette==0.37.2
tabulate==0.9.0
together==1.1.0
tomli==2.2.1
tqdm==4.67.1
typer==0.12.5
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
urllib3==2.5.0
uvicorn==0.30.1
vine==5.1.0
watchfiles==0.21.0
wcwidth==0.2.14
websockets==12.0
yarl==1.20.1
passlib[bcrypt]==1.7.4
PyJWT==2.10.1
email-validator==2.2.0
limits==3.13.0
slowapi==0.1.9
sentry-sdk==2.16.0
structlog==24.4.0
opentelemetry-sdk==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0
pip-audit==2.7.3

prometheus-client>=0.20,<1.0

